var name changes and expose more settings to initiator
What was originally called the "Buyer" is now the "Beneficiary"--the one who potentially benefits by receiving the DAI.

What was originally called the "Seller" is now the "Custodian"--the one responsible for making the final burn/release call.

Additionally, the initiator can set the beneficiaryDeposit to anything between 0 and the tradeAmount, and can also set the abortPunishment, to anything between 0 and beneficiaryDeposit.

Contract compiles, but has not yet been tested or analyzed rigorously for security.